CDPD-50545: Unified Analytics: Failed to remove partition pruning conjunct.
If a conjunct is being used in partition pruning, the filter isn't necessary
in the runtime nodes and can be removed. However, if there is a failure in
trying to remove the filter, it is not necessary to fail the query. There may
be a slight performance degradation which necessitates filing a warning, but
failing the query seems to be too harsh.

The behavior change here is because there is a specific case (CDPD-50499) where it
is hard to remove the filter. There is a filter expression like OR(A,B) that is
not matching OR(B,A), even though the expressions are equivalent. The OR expression
is being changed after the partition pruning expressions have been found.

The partition pruning will still happen so there should not be a major performance
degradation by leaving the filter in place. So it is better to leave the filter
and issue a warning rather than failing the query.

It should also be noted that the filter is not removed when being processed with
the Tez engine. The optimization only exists in the Unified Analytics code.

Change-Id: I1481fab63b8f46278309a35e259631cfdefaa6cb