DWX-14477: Throw SemanticException for INSERT into non-supported tables
Before this patch, HiveServer2 would send the execution request to
Impala's coordinator given the INSERT statement into AVRO or ORC tables
even though Impala's backend does not support it.

This patch generates a SemanticException for queries that attempt to
perform the INSERT operation on a table that is neither of text or
Parquet format so that HiveServer2 will not send the execution request
to Impala.

Moreover, with this patch, the error message generated by HiveServer2
due to such an operation could be displayed in Impala shell if a user
connects to HiveServer2 via Impala shell. Recall that before this patch,
a user was not able to see the error message returned from Impala's
backend even though the error message produced by Impala's backend could
be seen in beeline.

Testing:
 - Added a Hive q test to verify that HiveServer2 generates the error
   message for a query that inserts values into an ORC or an AVRO table.
 - Manually verified that HiveServer2 could produce a SemanticException
   for the INSERT operation on tables of unsupported formats and the
   error message could be shown in both beeline and Impala shell.

Change-Id: Ieb516ddbd09cab08c80f05aa06fe0ddfb1f6509e