CDPD-50380: Unified Analytics: Impala specific Compute Stats stmt fixed.
Only the basic parts of the COMPUTE STATS statement was working for
Impala.  The Impala specific variants were being dropped. For instance,
the "COMPUTE STATS <tbl> (column list)" was producing stats for all the
columns rather than only the column list that was specified in the sql
statement.

The Compute Stats statement for Impala was being derived from the
AST Tree for both Hive syntax and Impala syntax.  If the command
was originally in Impala syntax, there is no reason to rewrite
the query since it will drop all the Impala features added in the
command.

Change-Id: I86a1df6fee45d3de7e1fb58ff94e80add6a26ad0