DWX-15052: Support fetching delegation token for additional fs services
In PVC 1.5.1 we use the ozone filesystem in the form ofs://ozoneserviceid/... in several places,
also in a way where the default filesystem is not ozone hdfs, but the base hadoop hdfs service.
This means that two hdfs services are in use at the same time, but the query-coordinator is not
prepared for this, because it only fetches a delegation tokens for the default hdfs. In order
for the query-coordinator to be able to reach the ozone hdfs service, it needs a delegation token
to the ozone service as well. This commit introduces a new environment variable in the
query-coordinator entrypoint script which lists the additional hdfs fs services for which
a delegation token must be obtained in the script (in addition to the default hdfs).

Change-Id: I8d464b676df72bf6b9b9a51c1de41e4e7b7e0cd3