CDPD-56214: HIVE-27338: JDBC drivers are transitively inherited although declared optional (Stamatis Zampetakis reviewed by Alessandro Solimando)
1. Remove optional from dependencyManagement section since it has no
effect (MNG-5227).
2. Add explicitly the optional tag in every declaration of a JDBC
driver (except Derby) to:
* stop JDBC drivers propagation to other Hive modules and 3rd-party
projects;
* ensure license compliance for those drivers that are in Category-X
(https://www.apache.org/legal/resolved.html#optional)
3. Retain derby as mandatory (non-optional) in metastore modules to
avoid test failures in many modules that rely on it to run.
4. Remove (now) redundant JDBC driver exclusions from packaging module
(bin.xml).

Closes #4317

From the above changes (4) is not relevant and thus not backported.

Moreover the changes in the standalone-metastore module in the original
commit are not backported. In this branch the JDBC driver dependencies
are declared in the `dependencies` section not in
`dependencyManagement` thus the problem does not affect the
standalone-metastore module here.

Change-Id: I8c7f2fa7c245fb34519e11917f4a3896bd097958