DWX-15678: "Extract" functions were cast to incorrect function.
For functions like HOUR('03:05:14'), the existing code was resolving this to
use the HOUR(TIMESTAMP) function which was producing a NULL result. This instead
needs to resolve to HOUR(STRING)

This code change has a side effect of fixing an Impala issue. The function
YEAR('0009-03-01') resolves to YEAR(TIMESTAMP) in Impala which converts to the
incorrect value of NULL. This change picks the candidate YEAR(DATE) which picks
the correct value for YEAR('0009-03-01') and also anything that includes a time
like YEAR('2023-06-21 05:14:32')

Change-Id: I528e26a25816f1032fbd7bf58bc28657a2b62e83