CDPD-58509: Provide implemention for new Catalog.prioritizeLoad()
In Impala, the signature for Catalog.prioritizeLoad() changed upstream.
In order to avoid breaking the Hive build, a deprecated version
of prioritizeLoad was kept in downstream. This commit will change the
UA code to use the non-deprecated version so that the deprecated
version can be removed from the downstream codebase.

Change-Id: I337cfa634b36a750cec448bb9f0c3c4b88b56463