CDPD-58473: Refactor caller of checkTypeCompatibility
IMPALA-10173 changed the signature of the checkTypeCompatilibilty method. In order
to avoid breaking the build, the old deprecated implementation was left in the code.
This commit change the UA side so it points to the new version. After this push,
the deprecated code on the Impala side can be removed.

Change-Id: I0fc44a29d67991b2d9e3b8b12ef2a0cf114a5986