DWX-14022: TXNS table is not found during sys upgrade on 7.1.7 SP2
The table is created for 7.1.8 in hive-schema-3.1.3000.hive.sql:
https://github.infra.cloudera.com/CDH/hive/blob/9159f352c3b8b85bcb4e255b9dffc50b60761693/metastore/scripts/upgrade/hive/hive-schema-3.1.3000.hive.sql#L1350

But not in the 7.1.7 SP2 version of hive-schema-3.1.3000.hive.sql:
https://github.infra.cloudera.com/CDH/hive/blob/CDH-7.1.7.2000-305-tag/metastore/scripts/upgrade/hive/hive-schema-3.1.3000.hive.sql

Because of that, the upgrade `upgrade-3.1.3000.7.2.16.0-Update1-to-3.1.3000.7.2.16.0-Update2.hive` fails when it would try to select from the table.

Since 7.1.7 SP2 is already released, we need to work this around from CDWH, so I copied the creation of `TXNS` and `TXN_COMPONENTS` just before the creation of `TRANSACTIONS`, the same way it's done at other occurrences.

Change-Id: Ic4f586eded333051f604ff7493b9eaba7ea96ad7