CDPD-59995: Unified Analytics: Hive Java UDFs not loaded on HS2 initialization
On initialization of HS2, the Hive external UDFs are not getting
loaded in the Impala function map.

This is due to a call within ImpalaHelper where it tries to
"addHiveUDFs" before the ImpalaHelper object is fully loaded.

The HiveUDFs in the Impala function map will still get loaded at
initialization time after removing the statement here.  It will
happen immediately after they are loaded for the function map for
Tez functions.

Change-Id: I6acd93c9b8c9caaa9046b2c81a52592a3b518aa6