CDPD-41393: Apply Impala 'standardization' to all analytical expressions.
Previously we were only calling the 'standardize' function for a subset of
UDFs to be cautious. This caused some newly created functions to fail when used
in an analytical expression.

There are more test cases in place now versus when this was originally
developed and it seems the standardize method does not break any queries, so
it is ok to run it for all analytical expressions.

Change-Id: I296e4745524b66f28ab643849d4712bf413d66f9