CDPD-61980: Unified Analytics: support Impala syntax for create table statements.
This commit adds the ability to use create table syntax supported by Impala but
not Hive. For example, the "SORT BY" clause only exists in Impala (it is SORTED BY
in Hive).

This code uses its own Semantic Analyzer because the Table object does not exist
as it does in the other statements supported by the ImpalaSyntaxSemanticanalyzer.

In order to use the create database on CTAS statements, the "use <current db>"
stmt is run before running the "create table" syntax.

Change-Id: Ibcf3553f7efffdc8ef204aef8997b03fcb358cd1