HIVE-27873: Fix getOperationStatus and optimize fetch
This patch fixes a major performance issue fetching result from Impala.
The problem was that Impala does not set isHasResultSet during
getOperationStatus() calls, resulting in that RPC getting called and
logging a completion message for every row fetched. Optimizes the fetch
path to minimize conditional checks in the fast path.

Change-Id: Ifbcff2d9089b5df1330e6b5c6f7f50b3854176e3