CDPD-61191: Allow the value of TRuntimeFilterType to be numerical
This patch allows the value of TRuntimeFilterType to be numerical as
done by the Impala server.

In addition, we cleaned up some methods used in createQueryOptions().
Specifically, we revised some code in createDefaultQueryOptions() and
added more code comments. The signature of parseQueryOptions() was also
changed to match FeSupport#ParseQueryOptions(), which makes
createTestQueryOptions() easier to understand.

Change-Id: I0baaebc83a473ef5198cf9e49c2bc2e78e8892c2