CDPD-61296: Remove double quotation marks around Impala's query options
This patch removes the double quotation marks surrounding the value of
Impala's query options before they are sent to the Impala server to
establish the session. This aligns HiveServer2 with the classic Impala
because Impala's frontend is able to parse a value surrounded by the
double quotation marks if the SET statement is sent as a query to the
Impala server (v.s. configuration overlay).

Change-Id: Iceef44da739f59266632624b17bda8df9839a2ed