CDPD-36429: MySQL metastore schema upgrade from 1.2.1000 to 2.0.0 fails
The upgrade-1.2.1000-to-2.0.0.mysql.sql script cannot be processed by the MetastoreSchemaTool.
The tool doesn't understand the DELIMITER command so the switch from ';' to '$$' is not taken into account.
Due to this multiple statements are packed together incorrectly leading to syntax error when send to the MySQL server.

In BUG-107062, the content of 021-HIVE-7018.mysql.sql was incorrectly copied into upgrade-1.2.1000-to-2.0.0.mysql.sql leading to the problems mentioned above.

The procedure creation and the conditional removal of columns/keys is useless since the schema in version 1.2.1000 is guaranteed to have all of them.

Moreover, the copy of 021-HIVE-7018.mysql.sql into upgrade-1.2.1000-to-2.0.0.mysql.sql was partial.
The `CALL RM_LINKID` statement, which is supposed to run all procedures, was left out making the procedure declarations completely useless.

To solve the problem it suffices to include explicitly the ALTER ... DROP statements in the upgrade script without declaring procedures and conditional workflows.
This is inline with the unrolling of 021-HIVE-7018.mysql.sql into upgrade-1.2.0-to-2.0.0.mysql.sql in apache/master [1].

[1] https://github.com/apache/hive/blob/b6847ed38b7d32586ab22e224904867f159b510e/standalone-metastore/metastore-server/src/main/sql/mysql/upgrade-1.2.0-to-2.0.0.mysql.sql

Change-Id: Iec06a1e2b09d86d1f9007da010bdae00fb832eb6