CDPD-62550: Container initialization failed when running ITestOracle
The test fails since the Oracle container cannot be initialized within the specified threshold of 300 seconds.
In fact there is no problem with the container startup since the database starts just fine.
The problem is in the container readiness check that doesn't detect that the database is indeed ready.

CDPD-43236 changed the Oracle image but didn't update the container readiness check leading to the failure reported here.
Since ITestOracle is not run as part of the precommit tests the regression was left unnoticed.

Change-Id: Ib42dd8f7f26495f9275d3d61c8bdb04e6140650f