CDPD-62663: SCHEDULED_QUERIES tables are missing when upgrading from a metastore version before 3.1.3000 in MSSQL
The SCHEDULED_QUERIES and SCHEDULED_EXECUTIONS tables are present in a
fresh installation of the metastore at 3.1.3000 version (i.e.,
hive-schema-3.1.3000.mssql.sql) but are missing for those who are
upgrading from older versions cause their DDLs are not present in any
upgrade script.

The problem can be observed:
a) during a schema upgrade to a version after 3.1.3000.7.1.1.0;
b) at runtime when this feature is used in earlier versions.

For other DBMS, the tables are created when upgrading from 3.1.2000 to
3.1.3000, and the same should happen for MSSQL.

Change-Id: I03bd0d73184fe2627c04f4ff62138964a8a71141