CDPD-62661: Metastore schema upgrade from 1.2.1000-to-2.0.0 fails in MSSQL
The ALTER COLUMN statements inside the script attempt to update the
column type from nvarchar(128) to VARCHAR(1000) leading to the
misleading error below.
```
Error: The object 'COLUMNS_PK' is dependent on column 'COLUMN_NAME'.
```
The size increase from 128 to 1000 is safe and the legitimate purpose of
HIVE-11970 but the type change from NVARCHAR to VARCHAR has some
different implications causing the error above.

Changing the type was not intentional and since it was never tested
nobody realized there is a problem. The problem is not present in
apache/hive where the respective upgrade script uses the expected type
(nvarchar).

Change-Id: I474533049aaeb9a7485665fc883929e49ab9e945