CDPD-62662: Metastore schema upgrade from 3.1.1000 onwards fails in MSSQL
The ALTER statement command fails while attempting to ADD new column to
the table.
```
Error: Incorrect syntax near the keyword 'COLUMN'. (state=S0001,code=156)
```
The ADD COLUMN syntax that is used to do this is not valid in Microsoft
SQL Server. Using the COLUMN keyword in the ALTER statement is incorrect
and must be removed.

Change-Id: Id2869c8897bd4f1815446de6cdb93cf29fa86072