DWX-16853: Replace unnecessary notice files with latest ones
We build two images for the Hive repository, `hive` and `feng`. These two would each have its notice files which are required to be added by legal. The rest of the notice files are outdated and are probably originated from DWX where we originally built these images. As they are not necessary, we can safely remove them.

Change-Id: I781330f310f192461ea2c8f81347861b3a7bb4e0