HIVE-28018: Don't require HiveConf for JDBC Metadata calls
This patch eliminates the requirement to set hive.default.nulls.last in
HiveConf when calling DatabaseMetaData.nullsAreSorted(High/Low)
functions. The static default from HiveConf will be used if the HiveConf
or key is no present. When connecting to Hive databases, this is
normally not an issue since the server hiveconf will have this key.
However, the hs2 protocol is used outside of Hive (i.e. Impala) and the
server is not quaranteed to have hive keys.

Change-Id: Ifb64950a4b03120dcd28204e6cd9381458d7c8c7