DWX-17033: Fixed NPE for UA non BinaryPredicate join condition.
When the join condition in the UA code was not a binary predicate,
an NPE was being thrown.  There is an explicit check that checks
the left side and right side of the predicate for registering
information in the value transfer code. The predicate
that threw the exception was a "NOT" function which only had one
parameter.

The fix is to ensure that the join condition is a BinaryPredicate
when registering information in the value transfer code.

This has been tested via an e2e test on the Impala side which will
be committed after this has been pushed.

Change-Id: I8cb1c3ef473474664dbc985ca4300838cdea0656