CDPD-77710: Set up 'fileMetadataStats' when instantiating HdfsPartition
This patch sets up a non-null 'fileMetadataStats' when instantiating
HdfsPartition. Specifically, after IMPALA-13666, when calling the
constructor of HdfsPartition, we instantiate a FileMetadataStats using
a List of FileDescriptor's when they are available. This is the case
when we are instantiating an ImpalaHdfsPartition.

In addition, this patch makes the constructor of ImpalaBasicPartition
call the same constructor of the parent class as ImpalaHdfsPartition
does. This way Hive only needs one constructor of the parent class,
HdfsPartition, and hence we could safely remove other unused ones.

On a related note, it is fine to use -1L as the value of the field 'id'
when calling the constructor of HdfsPartition because the value of 'id'
is not actually used inside the constructor used by Hive.

This patch removes unused imports in ImpalaHdfsPartition.java and
ImpalaBasicPartition.java as well.

Testing:
 - Verified that with this patch, we could pass test_cmv() and
   test_cmv_sketches() that were disabled in part 2 of CDPD-77710
   in Impala.

Change-Id: I65e0f4b2862db6381436666e60e23eecae61f2f1