CDPD-81220: Do not install openssl10 in FENG docker image
In DWX-12786, we installed openssl compatibility library when producing
the docker image of FENG due to Impala being built on centos7 and the
image of FENG being based on ubi8, resulting in the SSL library that is
looked for to be older than expected on ubi8. Now that we have switched
to RedHat8 when building Impala, we should not need to install
openssl10 when building the FENG docker image.

Change-Id: Id9c0003e783766422fb9779f21685bd4ed8e26a6