CDPD-93756: Backport HIVE-13353 and the missing limiting option from HIVE-26580 (#2061)
This backport contains the full functionality of HIVE-13353 (SHOW COMPACTIONS should support filtering options).
Also contains the limiting functionality from HIVE-26580. In the scope of CDPD-74337 the ordering functionality got backported, but the limiting functionality was not.
It would be to useful to have the full functionality, so backporting the missing parts of HIVE-26580 here.

Please be aware, that upstream HIVE-13353 was implemented before the TxnHandler cleanup (HIVE-27481), but HIVE-27481 already got backported.
So the content of this commit is slightly different than the upstream commit for HIVE-13353.