CDPD-96850: CDPD-94381: Remove the Java code in the impala module
This patch removed the Java code in the impala module.

In impala/pom.xml we also removed all the dependencies.

The dependency on impala-frontend in packaging/pom.xml was removed as
well so that this jar file and those pulled in transitively will not be
included in the Hive package.

Change-Id: I6180f950639540f804a648b83c79161e5633cce7